remove migrations, add decimal for volume field

+++ scrape_api/bammens/copy_to_django.py
@@ -94,7 +94,7 @@ INSERT_TYPES = """
 INSERT INTO afvalcontainers_containertype
 SELECT id,
        data->>'name',
-       CAST("data"->>'volume' as INT),
+       CAST("data"->>'volume' as DECIMAL),
        CAST("data"->>'weight' as INT)
 FROM bammens_containertype_raw
 """
